Module barcoders::generators::image

Functionality for generating image representations of barcodes.

Each enum variant can be constructed via the standard constructor pattern or via a constructor method if you want default values.

For example:

use barcoders::generators::image::*;

// Specify your own struct fields.
let png = Image::PNG{height: 80, xdim: 1, rotation: Rotation::Zero};

// Or use the constructor for defaults.
let png = Image::png();

See the README for more examples.

Enums

Image
The image generator type.

Rotation
Possible rotation values for images.