Enum barcoders::sym::tf::TF

pub enum TF {
    Standard(Vec<u8>),
    Interleaved(Vec<u8>),
}

The 2-of-5 barcode type.

Variants

Standard

The standard 2-of-5 barcode type.

Interleaved

The interleaved 2-of-5 barcode type.

Methods

impl TF

fn interleaved(data: String) -> Result<TF>

Creates a new ITF barcode. If the length of the given data is odd, a checksum value will be computed and appended to the data for encoding.

Returns Result indicating parse success.

fn standard(data: String) -> Result<TF>

Creates a new STF barcode.

Returns Result indicating parse success.

fn checksum_digit(&self) -> Option<&u8>

Returns a Option reference to the checksum digit portion of the data. The Optional value is because Standard TF barcodes do not have a check digit.

fn encode(&self) -> Vec<u8>

Encodes the barcode. Returns a Vec of binary digits.